Remove unused components
- Fix problem with useRequest (create is called before abort)
- Separate out Table for RegionDetails
- Create hook useExtendedFiveW
