+++ src/hooks/useRequest.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useRef, useCallback } from 'react';
 import { isFalsyString } from '@togglecorp/fujs';
 import AbortController from 'abort-controller';
 
@@ -16,9 +16,43 @@ function useRequest<T>(
     schemaName: string,
     options: RequestInit = requestOption,
     preserveResponse = true,
+    // debug = false,
 ): [boolean, T | undefined] {
     const [response, setResponse] = useState<T>();
     const [pending, setPending] = useState(!!url);
+
+    const clientIdRef = useRef<number>(-1);
+    const pendingSetByRef = useRef<number>(-1);
+    const responseSetByRef = useRef<number>(-1);
+
+    const setPendingSafe = useCallback(
+        (value: boolean, clientId) => {
+            if (clientId >= pendingSetByRef.current) {
+                pendingSetByRef.current = clientId;
+                /*
+                if (debug) {
+                    console.warn('setting pending', value, 'by client', clientId);
+                }
+                */
+                setPending(value);
+            }
+        },
+        [],
+    );
+    const setResponseSafe = useCallback(
+        (value: T, clientId) => {
+            if (clientId >= responseSetByRef.current) {
+                responseSetByRef.current = clientId;
+                /*
+                if (debug) {
+                    console.warn('setting response', value, 'by client', clientId);
+                }
+                */
+                setResponse(value);
+            }
+        },
+        [],
+    );
 
     // NOTE: for warning only
     useEffect(
@@ -33,30 +67,40 @@ function useRequest<T>(
     useEffect(
         () => {
             if (isFalsyString(url)) {
-                setResponse(undefined);
-                setPending(false);
+                setResponseSafe(undefined, clientIdRef.current);
+                setPendingSafe(false, clientIdRef.current);
                 return () => {};
             }
-
             if (!preserveResponse) {
-                setResponse(undefined);
+                setResponseSafe(undefined, clientIdRef.current);
             }
-            setPending(true);
+
+            // console.warn('Creating new request', url);
+            clientIdRef.current += 1;
+
+            setPendingSafe(true, clientIdRef.current);
 
             const controller = new AbortController();
 
-            async function fetchResource(myUrl: string, myOptions: RequestInit | undefined) {
+            async function fetchResource(
+                myUrl: string,
+                myOptions: RequestInit | undefined,
+                clientId: number,
+            ) {
                 const { signal } = controller;
 
                 let res;
                 try {
                     res = await fetch(myUrl, { ...myOptions, signal });
                 } catch (e) {
-                    setPending(false);
+                    // console.warn('Errored fetch', url);
+                    setPendingSafe(false, clientId);
+
                     if (!signal.aborted) {
                         console.error(`An error occured while fetching ${myUrl}`, e);
+                        setResponseSafe(undefined, clientId);
                     } else {
-                        setResponse(undefined);
+                        // console.warn('Clearing response on network error');
                     }
                     return;
                 }
@@ -68,8 +112,9 @@ function useRequest<T>(
                         resBody = JSON.parse(resText);
                     }
                 } catch (e) {
-                    setResponse(undefined);
-                    setPending(false);
+                    // console.warn('Clearing response on parse error');
+                    setResponseSafe(undefined, clientId);
+                    setPendingSafe(false, clientId);
                     console.error(`An error occured while parsing data from ${myUrl}`, e);
                     return;
                 }
@@ -82,19 +127,18 @@ function useRequest<T>(
                             console.error(url, options.method, resBody, e.message);
                         }
                     }
-                    setResponse(resBody);
-                    setPending(false);
+                    setResponseSafe(resBody, clientId);
+                    setPendingSafe(false, clientId);
                 }
             }
 
-            fetchResource(url, options);
+            fetchResource(url, options, clientIdRef.current);
 
             return () => {
-                setPending(false);
                 controller.abort();
             };
         },
-        [url, options, schemaName, preserveResponse],
+        [url, options, schemaName, preserveResponse, setPendingSafe, setResponseSafe],
     );
 
     return [pending, response];
