--- src/views/Covid19/Stats/styles.css
@@ -1,57 +1,0 @@
-.stats {
-    .heading {
-        padding: var(--spacing-medium);
-        color: var(--color-accent);
-        font-weight: var(--font-weight-bold);
-    }
-
-    .content {
-        display: flex;
-        flex-wrap: wrap;
-
-        >div {
-            flex-basis: 50%;
-        }
-    }
-
-    .footer {
-        display: flex;
-        justify-content: space-between;
-    }
-}
-
-.stat-output {
-    padding: calc(var(--spacing-medium) - var(--spacing-extra-small)) calc(var(--spacing-medium) - var(--spacing-small));
-
-    .label {
-        padding: var(--spacing-extra-small) var(--spacing-small);
-        text-transform: capitalize;
-        font-size: var(--font-size-small);
-    }
-
-    .value {
-        padding: 0 var(--spacing-small);
-        color: var(--color-text-label);
-        font-size: var(--font-size-extra-large);
-        font-weight: var(--font-weight-bold);
-    }
-}
-
-.external-link {
-    display: flex;
-    align-items: center;
-    padding: calc(var(--spacing-medium) - var(--spacing-small));
-    text-decoration: none;
-    color: var(--color-primary);
-    font-size: var(--font-size-extra-small);
-    font-weight: var(--font-weight-bold);
-
-    .icon {
-        margin: var(--spacing-extra-small) var(--spacing-super-small);
-    }
-
-    .label {
-        padding: var(--spacing-extra-small) var(--spacing-small);
-        text-transform: capitalize;
-    }
-}
