--- src/views/Covid19/Tooltip/styles.css
@@ -1,29 +1,0 @@
-.list-item {
-    display: flex;
-    padding: var(--spacing-extra-small) var(--spacing-small) 0;
-
-    .count {
-        flex-shrink: 0;
-        margin-right: var(--spacing-small);
-    }
-    .label {
-        flex-grow: 1;
-    }
-}
-
-.tooltip {
-    display: flex;
-    flex-direction: column;
-    max-height: 300px;
-    overflow-y: auto;
-
-    .content {
-        display: flex;
-        flex-direction: column;
-
-        .sectors {
-            display: flex;
-            flex-direction: column;
-        }
-    }
-}
