--- src/views/Covid19/styles.css
@@ -1,109 +1,0 @@
-.covid19 {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-
-    .map-container {
-        flex-grow: 1;
-        font-family: inherit;
-    }
-
-    .map-style-config-container {
-        position: absolute;
-        bottom: var(--spacing-medium);
-        left: var(--spacing-medium);
-        border: var(--width-separator-thin) solid var(--color-separator);
-        background-color: var(--color-foreground);
-        padding: var(--spacing-medium);
-        width: 400px;
-
-        .separator {
-            margin: var(--spacing-large) 0;
-            border-bottom: var(--width-separator-thin) solid var(--color-separator);
-            width: 100%;
-        }
-
-        .input-item {
-            margin: var(--spacing-medium) 0;
-        }
-
-        .travel-time-input-container {
-            display: flex;
-            align-items: center;
-            margin: var(--spacing-medium) 0;
-        }
-
-        .hospitals {
-            display: flex;
-            flex-wrap: wrap;
-            padding: var(--spacing-extra-small) var(--spacing-small);
-
-            .button {
-                margin: var(--spacing-extra-small);
-                padding: var(--spacing-extra-small);
-                font-size: var(--font-size-small);
-            }
-        }
-
-        .abstract {
-            padding: var(--spacing-small);
-            font-size: var(--font-size-small);
-
-            .link {
-                margin: 0 var(--spacing-extra-small);
-                text-decoration: underline;
-                color: var(--color-link);
-            }
-        }
-    }
-
-    .legend-container {
-        display: flex;
-        position: absolute;
-        bottom: 10px;
-        left: calc(400px + 2 * var(--spacing-medium));
-        align-items: flex-end;
-        max-width: 512px;
-        max-height: 400px;
-
-        .legend {
-            margin: 0 var(--spacing-small);
-            border: var(--width-separator-thin) solid var(--color-separator);
-            background-color: var(--color-foreground);
-            padding: var(--spacing-medium);
-        }
-    }
-
-    .stats {
-        position: absolute;
-        top: var(--spacing-medium);
-        right: var(--spacing-medium);
-        border: var(--width-separator-thin) solid var(--color-separator);
-        background-color: var(--color-foreground);
-        padding: var(--spacing-medium);
-        width: 300px;
-    }
-
-    .print-mode-button {
-        position: absolute;
-        top: 10px;
-        left: 10px;
-        z-index: 1;
-    }
-
-    &.print-mode {
-        flex-grow: unset;
-        margin: auto;
-        width: 297mm!important;
-        height: 210mm!important;
-
-        .map-style-config-container {
-            display: none;
-        }
-
-        .legend-container {
-            bottom: calc(var(--height-print-bar) + 10px);
-            left: 10px;
-        }
-    }
-}
