+++ src/views/Dashboard/RegionDetails/Table/styles.css
@@ -1,0 +1,35 @@
+.table-actions {
+    display: flex;
+    align-items: flex-end;
+    flex-shrink: 0;
+    margin: var(--spacing-small) var(--spacing-medium);
+
+    >* {
+        &:not(:last-child) {
+            margin-right: var(--spacing-medium);
+        }
+    }
+}
+
+.table-container {
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    flex-grow: 1;
+    padding: var(--spacing-medium);
+    overflow: auto;
+
+    .backdrop {
+        z-index: 1;
+    }
+
+    .overflow-container {
+        flex-grow: 1;
+        z-index: 0;
+        overflow: auto;
+
+        .table {
+            min-width: 100%;
+        }
+    }
+}
