--- src/views/Glossary/styles.css
@@ -1,4 +1,0 @@
-.glossary {
-    position: relative;
-    height: var(--height-content);
-}
