--- src/views/Home/styles.css
@@ -1,8 +1,0 @@
-.home {
-    height: var(--height-content);
-
-    .message {
-        display: flex;
-        flex-direction: column;
-    }
-}
